HtmlBridge Phase 3 ratchet: de-list LayoutMetrics.cs — debt list now empty
LayoutMetrics.cs (2343 lines, ~3x the 750-line limit) was the last and largest
debt file. Decompose it into four cohesive sibling partials, leaving the core
box-metric accessors (clientWidth/offsetWidth/scrollWidth/..., offset/scroll-
parent resolution, bounding-client-rect, and the shared-geometry read-pass
cache) behind:

- LayoutMetrics.SvgAndZoom.cs (380): SVG geometry/text-metric resolution,
  element zoom / transform-scale resolution, border-box helpers, and the two
  [GeneratedRegex] transform-scale / text-path patterns.
- LayoutMetrics.Scrolling.cs (583): scrollIntoView option/argument parsing,
  element scroll-offset get/set with behaviour, scroll-event dispatch, visual-
  viewport scroll/scale, and programmatic-scrollability / overflow analysis.
- LayoutMetrics.ScrollGeometry.cs (400): scroll-container / fixed-position
  ancestor resolution, rendered-descendant enumeration, and the scrollIntoView
  physical-offset / scroll-coordinate / scroll-inset geometry conversions.
- LayoutMetrics.CssLength.cs (393): CSS <length> / calc()-style math evaluation
  against a viewport / containing-block basis, plus font-size / line-height
  reference resolution.

Extracted verbatim (line-range moves) — pure partial-class relocation, no
signature, accessibility, or logic change, so behaviour-identical by
construction. LayoutMetrics.cs drops 2343 -> 676 and its OversizedFileExemptions
entry is removed.

This empties the debt list: every HtmlBridge production file is now under 750
lines, and the guard runs as a pure ratchet (0 offenders / 0 exemptions).

Broiler.HtmlBridge.Dom builds clean; HtmlBridgeArchitectureGuardTests 14/14
green; geometry / scroll / SVG / viewport coverage
(LayoutGeometryCacheEquivalenceTests, LayoutGeometryCompletenessTests,
SharedLayoutGeometryTests, SharedGeometryZoomSizeTests, SharedScrollOverflowTests,
SvgDomAndCrossDocTests, SvgDomDynamicContentTests,
VisualViewportEventTargetBindingModuleTests, WindowScrollBindingModuleTests,
PositionAreaLiveGeometryTests, AnchorInsetLiveGeometryTests,
AbsPosInlineCbGeometryTests, 133 tests with the guard) green.
