+++ src/Broiler.HtmlBridge.Dom/DomBridge/LayoutMetrics.cs
@@ -673,1671 +673,4 @@ public sealed partial class DomBridge
         return (0, 0, 0, 0);
     }
 
-    private static bool IsSvgShapeElement(DomElement element)
-    {
-        var tag = element.TagName;
-        return string.Equals(tag, "rect", StringComparison.OrdinalIgnoreCase) ||
-               string.Equals(tag, "svg:rect", StringComparison.OrdinalIgnoreCase) ||
-               string.Equals(tag, "image", StringComparison.OrdinalIgnoreCase) ||
-               string.Equals(tag, "svg:image", StringComparison.OrdinalIgnoreCase) ||
-               string.Equals(tag, "foreignobject", StringComparison.OrdinalIgnoreCase) ||
-               string.Equals(tag, "svg:foreignobject", StringComparison.OrdinalIgnoreCase);
-    }
-
-    private static bool IsSvgViewportElement(DomElement element)
-    {
-        var tag = element.TagName;
-        return string.Equals(tag, "svg", StringComparison.OrdinalIgnoreCase) ||
-               string.Equals(tag, "svg:svg", StringComparison.OrdinalIgnoreCase);
-    }
-
-    private static bool IsSvgElement(DomElement element) =>
-        string.Equals(element.NamespaceUri, "http://www.w3.org/2000/svg", StringComparison.OrdinalIgnoreCase) ||
-        IsSvgViewportElement(element) ||
-        IsSvgShapeElement(element);
-
-    private double GetUsedZoomForElement(DomElement element)
-    {
-        var props = GetComputedProps(element);
-        var specifiedZoom = props.GetValueOrDefault("zoom");
-        var parentZoom = ParentEl(element) != null ? GetUsedZoomForElement(ParentEl(element)) : RootUsedZoomBase();
-        return ResolveSpecifiedZoom(specifiedZoom, parentZoom);
-    }
-
-    /// <summary>
-    /// The used-zoom base at the document root. Normally <c>1.0</c>; in native visual-viewport mode
-    /// (<see cref="NativeVisualViewport"/>) the document-root pinch-zoom scale is folded in here as a
-    /// root-level zoom, matching the extraction scale (<see cref="Broiler.Layout.Engine.NativeAnchorPlacement.VisualViewportScale"/>,
-    /// patch 0006) — so a scaled <c>BoxGeometry</c> divides back to unaffected <c>offset*</c> while
-    /// <c>getBoundingClientRect</c> stays scaled (this model treats pinch-zoom as a root zoom). Off the
-    /// native path this is <c>1.0</c>, so the DOM `zoom` bake continues to carry the factor unchanged.
-    /// </summary>
-    private double RootUsedZoomBase() =>
-        NativeVisualViewport && HasActiveVisualViewport() ? GetVisualViewportScale() : 1.0;
-
-    private static double ResolveSpecifiedZoom(string? specifiedZoom, double parentZoom)
-    {
-        if (string.IsNullOrWhiteSpace(specifiedZoom) ||
-            specifiedZoom.Equals("inherit", StringComparison.OrdinalIgnoreCase) ||
-            specifiedZoom.Equals("normal", StringComparison.OrdinalIgnoreCase))
-        {
-            return parentZoom;
-        }
-
-        if (double.TryParse(specifiedZoom, NumberStyles.Float,
-            CultureInfo.InvariantCulture, out var zoom) && zoom > 0)
-            return parentZoom * zoom;
-
-        return parentZoom;
-    }
-
-    private double GetTransformScale(DomElement element)
-    {
-        var transform = GetElementTransformValue(element);
-        if (string.IsNullOrWhiteSpace(transform))
-            return 1;
-
-        var match = GetTransformScaleRegex().Match(transform);
-        if (match.Success &&
-            double.TryParse(match.Groups["value"].Value, NumberStyles.Float,
-                CultureInfo.InvariantCulture, out double scale))
-        {
-            return scale;
-        }
-
-        return 1;
-    }
-
-    private string? GetElementTransformValue(DomElement element)
-    {
-        var props = GetComputedProps(element);
-        var transform = props.GetValueOrDefault("transform");
-        if (!string.IsNullOrWhiteSpace(transform) &&
-            !string.Equals(transform.Trim(), "none", StringComparison.OrdinalIgnoreCase))
-        {
-            return transform;
-        }
-
-        return TryGetAttribute(element, "transform", out var attributeTransform)
-            ? attributeTransform
-            : null;
-    }
-
-    private static bool IsSvgGroupElement(DomElement element)
-    {
-        var tag = element.TagName;
-        return string.Equals(tag, "g", StringComparison.OrdinalIgnoreCase) ||
-               string.Equals(tag, "svg:g", StringComparison.OrdinalIgnoreCase);
-    }
-
-    private static bool IsSvgTextContentElement(DomElement element)
-    {
-        var tag = element.TagName;
-        return string.Equals(tag, "text", StringComparison.OrdinalIgnoreCase) ||
-               string.Equals(tag, "svg:text", StringComparison.OrdinalIgnoreCase) ||
-               string.Equals(tag, "tspan", StringComparison.OrdinalIgnoreCase) ||
-               string.Equals(tag, "svg:tspan", StringComparison.OrdinalIgnoreCase) ||
-               string.Equals(tag, "textpath", StringComparison.OrdinalIgnoreCase) ||
-               string.Equals(tag, "svg:textpath", StringComparison.OrdinalIgnoreCase);
-    }
-
-    private bool TryGetSvgChildrenUnionRect(DomElement element,
-        out (double Left, double Top, double Width, double Height) rect)
-    {
-        var found = false;
-        var minLeft = 0d;
-        var minTop = 0d;
-        var maxRight = 0d;
-        var maxBottom = 0d;
-
-        foreach (var child in ChildElements(element))
-        {
-            if (IsText(child) || child.TagName.StartsWith('#'))
-                continue;
-
-            var (Left, Top, Width, Height) = GetHitTestRectForElement(child);
-            if (Width <= 0 || Height <= 0)
-                continue;
-
-            if (!found)
-            {
-                found = true;
-                minLeft = Left;
-                minTop = Top;
-                maxRight = Left + Width;
-                maxBottom = Top + Height;
-                continue;
-            }
-
-            minLeft = Math.Min(minLeft, Left);
-            minTop = Math.Min(minTop, Top);
-            maxRight = Math.Max(maxRight, Left + Width);
-            maxBottom = Math.Max(maxBottom, Top + Height);
-        }
-
-        rect = found
-            ? (minLeft, minTop, Math.Max(0, maxRight - minLeft), Math.Max(0, maxBottom - minTop))
-            : (0, 0, 0, 0);
-        return found;
-    }
-
-    private bool TryGetSvgTextHitTestRect(DomElement element,
-        out (double Left, double Top, double Width, double Height) rect)
-    {
-        var text = GetDirectTextContent(element);
-        if (string.IsNullOrWhiteSpace(text))
-        {
-            rect = (0, 0, 0, 0);
-            return false;
-        }
-
-        var fontSize = ResolveFontSizeForElement(element);
-        if (fontSize <= 0)
-            fontSize = 16;
-
-        var width = text
-            .Replace("\r", string.Empty)
-            .Split('\n')
-            .Select(line => line.Trim())
-            .Where(line => line.Length > 0)
-            .DefaultIfEmpty(string.Empty)
-            .Max(line => line.Length) * fontSize * 0.6;
-        if (width <= 0)
-        {
-            rect = (0, 0, 0, 0);
-            return false;
-        }
-
-        var baselineX = ResolveSvgTextCoordinate(element, "x");
-        var baselineY = ResolveSvgTextCoordinate(element, "y");
-        if (IsSvgTextPathElement(element) &&
-            TryResolveSvgTextPathStart(element, out var pathStart))
-        {
-            if (!HasOwnSvgCoordinate(element, "x"))
-                baselineX = pathStart.X;
-            if (!HasOwnSvgCoordinate(element, "y"))
-                baselineY = pathStart.Y;
-        }
-
-        var viewport = FindNearestSvgViewportAncestor(element);
-        if (viewport != null)
-        {
-            var (Left, Top, Width, Height) = ComputeRenderedRect(viewport);
-            baselineX += Left;
-            baselineY += Top;
-        }
-
-        rect = (baselineX, baselineY - fontSize, width, fontSize);
-        return true;
-    }
-
-    private static string GetDirectTextContent(DomElement element)
-    {
-        // RF-BRIDGE-1c Phase F (F3c part 2d): a node's direct text is its text-node children.
-        var sb = new StringBuilder();
-        foreach (var child in element.ChildNodes)
-        {
-            if (IsText(child) && !string.IsNullOrWhiteSpace(BridgeText(child)))
-                sb.Append(BridgeText(child));
-        }
-
-        return sb.ToString();
-    }
-
-    private double ResolveSvgTextCoordinate(DomElement element, string attributeName)
-    {
-        for (var current = element; current != null; current = ParentEl(current))
-        {
-            if (!IsSvgTextContentElement(current))
-                continue;
-
-            if (TryGetAttribute(current, attributeName, out var rawValue))
-            {
-                var percentageBasis = ResolveContainingBlockReferenceLength(
-                    current,
-                    vertical: string.Equals(attributeName, "y", StringComparison.OrdinalIgnoreCase));
-                var resolved = ParseCssLengthToPixelsWithViewport(rawValue, current, percentageBasis: percentageBasis);
-                if (resolved > 0 || string.Equals(rawValue?.Trim(), "0", StringComparison.Ordinal))
-                    return resolved;
-
-                var scalar = rawValue?
-                    .Split([' ', '\t', '\r', '\n', ','], StringSplitOptions.RemoveEmptyEntries)
-                    .FirstOrDefault();
-                if (double.TryParse(
-                    scalar,
-                    NumberStyles.Float,
-                    CultureInfo.InvariantCulture,
-                    out var numericValue))
-                {
-                    return numericValue;
-                }
-            }
-
-            if (IsSvgTextPathElement(current) &&
-                TryResolveSvgTextPathStart(current, out var pathStart))
-            {
-                return string.Equals(attributeName, "y", StringComparison.OrdinalIgnoreCase)
-                    ? pathStart.Y
-                    : pathStart.X;
-            }
-        }
-
-        return 0;
-    }
-
-    private static bool HasOwnSvgCoordinate(DomElement element, string attributeName) =>
-        TryGetAttribute(element, attributeName, out var rawValue) &&
-        !string.IsNullOrWhiteSpace(rawValue);
-
-    private bool TryResolveSvgTextPathStart(DomElement element, out (double X, double Y) point)
-    {
-        point = default;
-        if (!TryGetAttribute(element, "href", out var href) &&
-            !TryGetAttribute(element, "xlink:href", out href))
-        {
-            return false;
-        }
-
-        href = href?.Trim();
-        if (string.IsNullOrWhiteSpace(href) || !href.StartsWith('#'))
-            return false;
-
-        var documentElement = GetOwningDocumentElement(element);
-        var referencedPath = documentElement != null
-            ? FindInTree(documentElement, candidate => string.Equals(candidate.Id, href[1..], StringComparison.Ordinal))
-            : null;
-        if (referencedPath == null ||
-            !TryGetAttribute(referencedPath, "d", out var pathData) ||
-            string.IsNullOrWhiteSpace(pathData))
-        {
-            return false;
-        }
-
-        var moveMatch = TryResolveSvgTextPathStartRegex().Match(pathData);
-        if (!moveMatch.Success ||
-            !double.TryParse(moveMatch.Groups["x"].Value, NumberStyles.Float,
-                CultureInfo.InvariantCulture, out var x) ||
-            !double.TryParse(moveMatch.Groups["y"].Value, NumberStyles.Float,
-                CultureInfo.InvariantCulture, out var y))
-        {
-            return false;
-        }
-
-        point = (x, y);
-        return true;
-    }
-
-    private static bool IsSvgTextPathElement(DomElement element)
-    {
-        var tag = element.TagName;
-        return string.Equals(tag, "textpath", StringComparison.OrdinalIgnoreCase) ||
-               string.Equals(tag, "svg:textpath", StringComparison.OrdinalIgnoreCase);
-    }
-
-    private static DomElement? FindNearestSvgViewportAncestor(DomElement element)
-    {
-        for (var current = ParentEl(element); current != null; current = ParentEl(current))
-        {
-            if (IsSvgViewportElement(current))
-                return current;
-        }
-
-        return null;
-    }
-
-    private double GetBorderBoxWidth(Dictionary<string, string> props, DomElement? element = null)
-    {
-        var containingBlockWidth = element != null ? ResolveContainingBlockReferenceLength(element, vertical: false) : (double?)null;
-        return ParseCssLengthToPixelsWithViewport(props.GetValueOrDefault("width"), element, percentageBasis: containingBlockWidth)
-             + ParseCssLengthToPixelsWithViewport(props.GetValueOrDefault("padding-left"), element, percentageBasis: containingBlockWidth)
-             + ParseCssLengthToPixelsWithViewport(props.GetValueOrDefault("padding-right"), element, percentageBasis: containingBlockWidth)
-             + ParseCssLengthToPixelsWithViewport(props.GetValueOrDefault("border-left-width"), element)
-             + ParseCssLengthToPixelsWithViewport(props.GetValueOrDefault("border-right-width"), element);
-    }
-
-    private double GetBorderBoxHeight(Dictionary<string, string> props, DomElement? element = null)
-    {
-        var containingBlockWidth = element != null ? ResolveContainingBlockReferenceLength(element, vertical: false) : (double?)null;
-        var containingBlockHeight = element != null ? ResolveContainingBlockReferenceLength(element, vertical: true) : (double?)null;
-        return ParseCssLengthToPixelsWithViewport(props.GetValueOrDefault("height"), element, percentageBasis: containingBlockHeight)
-             + ParseCssLengthToPixelsWithViewport(props.GetValueOrDefault("padding-top"), element, percentageBasis: containingBlockWidth)
-             + ParseCssLengthToPixelsWithViewport(props.GetValueOrDefault("padding-bottom"), element, percentageBasis: containingBlockWidth)
-             + ParseCssLengthToPixelsWithViewport(props.GetValueOrDefault("border-top-width"), element)
-             + ParseCssLengthToPixelsWithViewport(props.GetValueOrDefault("border-bottom-width"), element);
-    }
-
-    private void ScrollElementIntoView(DomElement element,
-        string? block = null, string? inline = null, string? behavior = null)
-    {
-        var current = element;
-        for (var i = 0; i < MaxScrollContinuationDepth && current != null; i++)
-        {
-            var scrollContainer = FindScrollContainer(current) ?? GetOwningDocumentElement(current);
-            if (scrollContainer == null)
-                return;
-
-            if (IsDocumentElement(scrollContainer) && HasFixedPositionInDocument(current, scrollContainer))
-            {
-                if (HasActiveVisualViewport())
-                {
-                    ScrollFixedElementIntoVisualViewport(element, scrollContainer, block, inline);
-                    current = GetOuterFrameElement(scrollContainer);
-                    continue;
-                }
-
-                current = GetOuterFrameElement(scrollContainer);
-                continue;
-            }
-
-            var (horizontalAlignment, verticalAlignment) = ResolvePhysicalScrollIntoViewAlignments(
-                scrollContainer,
-                block,
-                inline);
-            var scrollTop = ResolveScrollIntoViewOffset(element, scrollContainer, vertical: true, alignment: verticalAlignment);
-            var scrollLeft = ResolveScrollIntoViewOffset(element, scrollContainer, vertical: false, alignment: horizontalAlignment);
-
-            SetElementScrollOffsetsWithBehavior(scrollContainer, scrollLeft, scrollTop, clamp: true, behavior: behavior);
-
-            var next = GetOuterScrollContinuationElement(scrollContainer);
-            if (next == null || ReferenceEquals(next, current))
-                return;
-
-            current = next;
-        }
-    }
-
-    private (string Block, string Inline, string? Behavior) GetScrollIntoViewOptions(in Arguments args)
-    {
-        const string defaultBlock = "start";
-        const string defaultInline = "nearest";
-
-        if (args.Length == 0)
-            return (defaultBlock, "start-if-needed", null);
-
-        var first = args[0];
-        if (first is JSObject options)
-        {
-            return (
-                NormalizeScrollIntoViewAlignment(GetOptionalStringOption(options, "block"), defaultBlock),
-                NormalizeScrollIntoViewAlignment(GetOptionalStringOption(options, "inline"), defaultInline),
-                GetOptionalScrollBehavior(options));
-        }
-
-        if (first.IsBoolean)
-        {
-            return first.BooleanValue
-                ? (defaultBlock, defaultInline, null)
-                : ("end", defaultInline, null);
-        }
-
-        return (defaultBlock, defaultInline, null);
-    }
-
-    private (double? Left, double? Top, string? Behavior) GetScrollArguments(in Arguments args)
-    {
-        if (args.Length == 0)
-            return (null, null, null);
-
-        if (args[0] is JSObject options)
-        {
-            return (
-                GetOptionalScrollCoordinate(options, "left"),
-                GetOptionalScrollCoordinate(options, "top"),
-                GetOptionalScrollBehavior(options));
-        }
-
-        return (args.Length > 0 ? args[0].DoubleValue : null, args.Length > 1 ? args[1].DoubleValue : null, null);
-    }
-
-    private static double? GetOptionalScrollCoordinate(JSObject options, string propertyName)
-    {
-        var value = options[(KeyString)propertyName];
-        return value == null || value.IsUndefined || value.IsNull ? null : value.DoubleValue;
-    }
-
-    private static string? GetOptionalScrollBehavior(JSObject options)
-    {
-        var value = options[(KeyString)"behavior"];
-        if (value == null || value.IsUndefined || value.IsNull)
-            return null;
-
-        var behavior = value.ToString();
-        return string.IsNullOrWhiteSpace(behavior) ? null : behavior;
-    }
-
-    private static string? GetOptionalStringOption(JSObject options, string propertyName)
-    {
-        var value = options[(KeyString)propertyName];
-        if (value == null || value.IsUndefined || value.IsNull)
-            return null;
-
-        var text = value.ToString();
-        return string.IsNullOrWhiteSpace(text) ? null : text;
-    }
-
-    private static string NormalizeScrollIntoViewAlignment(string? value, string fallback)
-    {
-        if (string.IsNullOrWhiteSpace(value))
-            return fallback;
-
-        var normalized = value.Trim().ToLowerInvariant();
-        return normalized is "start" or "center" or "end" or "nearest" or "start-if-needed" or "end-if-needed"
-            ? normalized
-            : fallback;
-    }
-
-    private (string Horizontal, string Vertical) ResolvePhysicalScrollIntoViewAlignments(
-        DomElement scrollContainer, string? block, string? inline)
-    {
-        var props = GetComputedProps(scrollContainer);
-        var writingMode = props.GetValueOrDefault("writing-mode")?.Trim().ToLowerInvariant();
-        var direction = props.GetValueOrDefault("direction");
-        bool isVerticalWritingMode = IsVerticalWritingMode(writingMode);
-        bool isRtl = string.Equals(direction, "rtl", StringComparison.OrdinalIgnoreCase);
-
-        var horizontal = ResolvePhysicalAxisAlignment(
-            alignment: isVerticalWritingMode ? block : inline,
-            startMapsToPhysicalStart: !isVerticalWritingMode
-                ? !isRtl
-                : writingMode?.EndsWith("-rl", StringComparison.Ordinal) != true);
-        var vertical = ResolvePhysicalAxisAlignment(
-            alignment: isVerticalWritingMode ? inline : block,
-            startMapsToPhysicalStart: !isVerticalWritingMode || !isRtl);
-        return (horizontal, vertical);
-    }
-
-    private static string ResolvePhysicalAxisAlignment(string? alignment, bool startMapsToPhysicalStart)
-    {
-        var normalized = NormalizeScrollIntoViewAlignment(alignment, "start");
-        if (normalized is "center" or "nearest" || startMapsToPhysicalStart)
-            return normalized;
-
-        return normalized switch
-        {
-            "start" => "end",
-            "end" => "start",
-            "start-if-needed" => "end-if-needed",
-            "end-if-needed" => "start-if-needed",
-            _ => normalized
-        };
-    }
-
-    private double GetElementScrollOffset(DomElement element, bool vertical)
-    {
-        if (!CanProgrammaticallyScroll(element, vertical))
-            return 0;
-
-        return TryGetStoredScrollOffset(element, vertical, out var scrollOffset)
-            ? scrollOffset
-            : 0;
-    }
-
-    private (double Left, double Top) ResolveElementScrollOffsets(DomElement element, double? left = null, double? top = null, bool relative = false, bool clamp = true)
-    {
-        var currentLeft = GetElementScrollOffset(element, vertical: false);
-        var currentTop = GetElementScrollOffset(element, vertical: true);
-
-        var nextLeft = left.HasValue ? (relative ? currentLeft + left.Value : left.Value) : currentLeft;
-        var nextTop = top.HasValue ? (relative ? currentTop + top.Value : top.Value) : currentTop;
-
-        if (!CanProgrammaticallyScroll(element, vertical: false))
-            nextLeft = 0;
-        if (!CanProgrammaticallyScroll(element, vertical: true))
-            nextTop = 0;
-
-        if (clamp)
-        {
-            var (minLeft, maxLeft, minTop, maxTop) = GetScrollBounds(element);
-            nextLeft = Math.Clamp(nextLeft, minLeft, maxLeft);
-            nextTop = Math.Clamp(nextTop, minTop, maxTop);
-        }
-
-        return (nextLeft, nextTop);
-    }
-
-    private void SetElementScrollOffsetsWithBehavior(DomElement element,
-        double? left = null, double? top = null,
-        bool relative = false, bool clamp = true,
-        string? behavior = null)
-    {
-        var trackVisualViewport = ReferenceEquals(element, DocumentElement);
-        var previousVisualPageLeft = trackVisualViewport ? GetVisualViewportPageOffset(vertical: false) : 0;
-        var previousVisualPageTop = trackVisualViewport ? GetVisualViewportPageOffset(vertical: true) : 0;
-        var previousLeft = GetElementScrollOffset(element, vertical: false);
-        var previousTop = GetElementScrollOffset(element, vertical: true);
-        var (targetLeft, targetTop) = ResolveElementScrollOffsets(element, left, top, relative, clamp);
-        var hadActiveSmoothScroll = _smoothScrollTokens.ContainsKey(element);
-        var effectiveBehavior = ResolveScrollBehavior(element, behavior);
-        if (hadActiveSmoothScroll && NormalizeScrollBehavior(behavior) != "smooth")
-            effectiveBehavior = "instant";
-        CancelSmoothScroll(element);
-
-        if (string.Equals(effectiveBehavior, "smooth", StringComparison.OrdinalIgnoreCase))
-        {
-            var token = Interlocked.Increment(ref _smoothScrollTokenCounter);
-            _smoothScrollTokens[element] = token;
-            QueueFrameAction(() =>
-            {
-                if (_smoothScrollTokens.TryGetValue(element, out var activeToken) && activeToken == token)
-                {
-                    var queuedPreviousLeft = GetElementScrollOffset(element, vertical: false);
-                    var queuedPreviousTop = GetElementScrollOffset(element, vertical: true);
-                    var queuedPreviousVisualPageLeft = trackVisualViewport ? GetVisualViewportPageOffset(vertical: false) : 0;
-                    var queuedPreviousVisualPageTop = trackVisualViewport ? GetVisualViewportPageOffset(vertical: true) : 0;
-                    GetElementRuntimeState(element).Scroll.Left.Set(targetLeft);
-                    GetElementRuntimeState(element).Scroll.Top.Set(targetTop);
-                    NotifyVisualViewportScrollIfNeeded(queuedPreviousVisualPageLeft, queuedPreviousVisualPageTop, trackVisualViewport);
-                    DispatchScrollEventIfNeeded(element, queuedPreviousLeft, queuedPreviousTop);
-                    DispatchScrollEndEventIfNeeded(element, queuedPreviousLeft, queuedPreviousTop);
-                    _smoothScrollTokens.TryRemove(element, out _);
-                }
-            });
-
-            // Approximate smooth scrolling with a visible intermediate frame before
-            // finishing on the next queued frame.
-            GetElementRuntimeState(element).Scroll.Left.Set(previousLeft + ((targetLeft - previousLeft) / 2.0));
-            GetElementRuntimeState(element).Scroll.Top.Set(previousTop + ((targetTop - previousTop) / 2.0));
-            NotifyVisualViewportScrollIfNeeded(previousVisualPageLeft, previousVisualPageTop, trackVisualViewport);
-            DispatchScrollEventIfNeeded(element, previousLeft, previousTop);
-            return;
-        }
-
-        GetElementRuntimeState(element).Scroll.Left.Set(targetLeft);
-        GetElementRuntimeState(element).Scroll.Top.Set(targetTop);
-        NotifyVisualViewportScrollIfNeeded(previousVisualPageLeft, previousVisualPageTop, trackVisualViewport);
-        DispatchScrollEventIfNeeded(element, previousLeft, previousTop);
-        DispatchScrollEndEventIfNeeded(element, previousLeft, previousTop);
-    }
-
-    private void QueueFrameAction(Action callback) => _eventLoop.QueueFrameAction(callback);
-
-    private void CancelSmoothScroll(DomElement element) => _smoothScrollTokens.TryRemove(element, out _);
-
-    private void DispatchScrollEventIfNeeded(DomElement element, double previousLeft, double previousTop)
-    {
-        if (AreClose(previousLeft, GetElementScrollOffset(element, vertical: false)) &&
-            AreClose(previousTop, GetElementScrollOffset(element, vertical: true)))
-            return;
-
-        DispatchElementEvent(element, "scroll");
-    }
-
-    private void DispatchScrollEndEventIfNeeded(DomElement element, double previousLeft, double previousTop)
-    {
-        if (AreClose(previousLeft, GetElementScrollOffset(element, vertical: false)) &&
-            AreClose(previousTop, GetElementScrollOffset(element, vertical: true)))
-            return;
-
-        DispatchElementEvent(element, "scrollend");
-    }
-
-    private void DispatchElementEvent(DomElement element, string eventType)
-    {
-        var evt = new JSObject();
-        evt.FastAddValue((KeyString)"type", new JSString(eventType), JSPropertyAttributes.EnumerableConfigurableValue);
-        evt.FastAddValue((KeyString)"bubbles", JSBoolean.False, JSPropertyAttributes.EnumerableConfigurableValue);
-        DispatchEventOnElement(element, evt);
-    }
-
-    private string ResolveScrollBehavior(DomElement element, string? requestedBehavior)
-    {
-        var normalizedRequested = NormalizeScrollBehavior(requestedBehavior);
-        if (normalizedRequested == "instant" || normalizedRequested == "smooth")
-            return normalizedRequested;
-
-        var props = GetComputedProps(element);
-        return NormalizeScrollBehavior(props.GetValueOrDefault("scroll-behavior")) == "smooth"
-            ? "smooth"
-            : "instant";
-    }
-
-    private static string NormalizeScrollBehavior(string? behavior)
-    {
-        if (string.IsNullOrWhiteSpace(behavior))
-            return "auto";
-
-        var normalized = behavior.Trim().ToLowerInvariant();
-        return normalized is "instant" or "smooth" ? normalized : "auto";
-    }
-
-    private bool HasActiveVisualViewport() => GetVisualViewportScale() > 1.0001;
-
-    private double GetVisualViewportScale() => _visualViewportScale > 1 ? _visualViewportScale : 1;
-
-    private double GetVisualViewportWidth() => _viewportWidth / GetVisualViewportScale();
-
-    private double GetVisualViewportHeight() => _viewportHeight / GetVisualViewportScale();
-
-    private double GetVisualViewportPageOffset(bool vertical)
-    {
-        var layoutOffset = GetElementScrollOffset(DocumentElement, vertical);
-        return layoutOffset + GetVisualViewportExtraOffset(vertical);
-    }
-
-    private void SetVisualViewportScale(double scale)
-    {
-        _visualViewportScale = double.IsFinite(scale) && scale > 1 ? scale : 1;
-        ClampVisualViewportOffsets();
-    }
-
-    private void ScrollFixedElementIntoVisualViewport(DomElement element, DomElement scrollContainer,
-        string? block, string? inline)
-    {
-        var targetTop = ResolveScrollIntoViewOffset(
-            element,
-            scrollContainer,
-            vertical: true,
-            alignment: block,
-            viewportSizeOverride: GetVisualViewportHeight(),
-            currentScrollOverride: GetVisualViewportPageOffset(vertical: true),
-            offsetOverride: GetElementScrollOffset(scrollContainer, vertical: true) +
-                OffsetWithinAncestorForFixedPreferShared(element, scrollContainer, vertical: true),
-            coordinateSpaceIsPhysical: true);
-        var targetLeft = ResolveScrollIntoViewOffset(
-            element,
-            scrollContainer,
-            vertical: false,
-            alignment: inline,
-            viewportSizeOverride: GetVisualViewportWidth(),
-            currentScrollOverride: GetVisualViewportPageOffset(vertical: false),
-            offsetOverride: GetElementScrollOffset(scrollContainer, vertical: false) +
-                OffsetWithinAncestorForFixedPreferShared(element, scrollContainer, vertical: false),
-            coordinateSpaceIsPhysical: true);
-        SetVisualViewportPageOffsets(left: targetLeft, top: targetTop);
-    }
-
-    private void SetVisualViewportPageOffsets(double? left = null, double? top = null)
-    {
-        var oldPageLeft = GetVisualViewportPageOffset(vertical: false);
-        var oldPageTop = GetVisualViewportPageOffset(vertical: true);
-        var layoutLeft = GetElementScrollOffset(DocumentElement, vertical: false);
-        var layoutTop = GetElementScrollOffset(DocumentElement, vertical: true);
-
-        if (left.HasValue)
-        {
-            _visualViewportPageLeftOffset = Math.Clamp(
-                left.Value - layoutLeft,
-                0,
-                GetVisualViewportMaxExtraOffset(vertical: false));
-        }
-
-        if (top.HasValue)
-        {
-            _visualViewportPageTopOffset = Math.Clamp(
-                top.Value - layoutTop,
-                0,
-                GetVisualViewportMaxExtraOffset(vertical: true));
-        }
-
-        if (!AreClose(oldPageLeft, GetVisualViewportPageOffset(vertical: false)) ||
-            !AreClose(oldPageTop, GetVisualViewportPageOffset(vertical: true)))
-        {
-            DispatchVisualViewportScrollEvent();
-        }
-    }
-
-    private void ClampVisualViewportOffsets()
-    {
-        _visualViewportPageLeftOffset = Math.Clamp(_visualViewportPageLeftOffset, 0, GetVisualViewportMaxExtraOffset(vertical: false));
-        _visualViewportPageTopOffset = Math.Clamp(_visualViewportPageTopOffset, 0, GetVisualViewportMaxExtraOffset(vertical: true));
-    }
-
-    private double GetVisualViewportExtraOffset(bool vertical) =>
-        vertical ? _visualViewportPageTopOffset : _visualViewportPageLeftOffset;
-
-    private double GetVisualViewportMaxExtraOffset(bool vertical)
-    {
-        if (!HasActiveVisualViewport())
-            return 0;
-
-        var layoutSize = vertical ? _viewportHeight : _viewportWidth;
-        var visualSize = vertical ? GetVisualViewportHeight() : GetVisualViewportWidth();
-        return Math.Max(0, layoutSize - visualSize);
-    }
-
-    private void DispatchVisualViewportScrollEvent()
-    {
-        if (_visualViewportJSObject == null || _eventTargets.VisualViewportScrollListeners.Count == 0)
-            return;
-
-        var evt = new JSObject();
-        evt.FastAddValue((KeyString)"type", new JSString("scroll"), JSPropertyAttributes.EnumerableConfigurableValue);
-        evt.FastAddValue((KeyString)"target", _visualViewportJSObject, JSPropertyAttributes.EnumerableConfigurableValue);
-        evt.FastAddValue((KeyString)"currentTarget", _visualViewportJSObject, JSPropertyAttributes.EnumerableConfigurableValue);
-
-        foreach (var listener in _eventTargets.VisualViewportScrollListeners.ToList())
-        {
-            try
-            {
-                listener.InvokeFunction(new Arguments(listener, evt));
-            }
-            catch (Exception ex)
-            {
-                RenderLogger.LogWarning(LogCategory.JavaScript, "DomBridge.visualViewport", $"Visual viewport listener error: {ex.Message}", ex);
-            }
-        }
-    }
-
-    private static bool AreClose(double left, double right) => Math.Abs(left - right) < 0.0001;
-
-    private void NotifyVisualViewportScrollIfNeeded(double previousPageLeft, double previousPageTop, bool trackVisualViewport)
-    {
-        if (!trackVisualViewport)
-            return;
-
-        if (!AreClose(previousPageLeft, GetVisualViewportPageOffset(vertical: false)) ||
-            !AreClose(previousPageTop, GetVisualViewportPageOffset(vertical: true)))
-        {
-            DispatchVisualViewportScrollEvent();
-        }
-    }
-
-    private bool CanProgrammaticallyScroll(DomElement element, bool vertical)
-    {
-        if (IsDocumentElement(element) ||
-            IsViewportBodyElement(element, GetOwningDocumentElement(element)))
-        {
-            return CanProgrammaticallyScrollRoot(element, vertical);
-        }
-
-        if (IsSelectListBox(element))
-            return CanProgrammaticallyScrollSelectListBox(element, vertical);
-
-        var props = GetComputedProps(element);
-        var axisValue = GetOverflowAxisValue(props, vertical);
-
-        return EnablesScrollingBox(axisValue);
-    }
-
-    private bool CanProgrammaticallyScrollRoot(DomElement rootElement, bool vertical)
-    {
-        var documentElement = GetOwningDocumentElement(rootElement);
-        var htmlOverflow = GetOverflowAxisValue(GetComputedProps(documentElement), vertical);
-        var body = FindBodyElement(documentElement);
-        var bodyOverflow = body != null ? GetOverflowAxisValue(GetComputedProps(body), vertical) : null;
-
-        if (DisablesRootScrolling(htmlOverflow) || DisablesRootScrolling(bodyOverflow))
-            return false;
-
-        return true;
-    }
-
-    private static string? GetOverflowAxisValue(Dictionary<string, string> props, bool vertical)
-    {
-        var axisValue = props.GetValueOrDefault(vertical ? "overflow-y" : "overflow-x");
-        if (string.IsNullOrWhiteSpace(axisValue))
-            axisValue = props.GetValueOrDefault("overflow");
-        return axisValue;
-    }
-
-    private static bool DisablesRootScrolling(string? overflowValue)
-    {
-        if (string.IsNullOrWhiteSpace(overflowValue))
-            return false;
-
-        var normalized = overflowValue.Trim().ToLowerInvariant();
-        return normalized.Contains("hidden") || normalized.Contains("clip");
-    }
-
-    private static bool EnablesScrollingBox(string? overflowValue)
-    {
-        if (string.IsNullOrWhiteSpace(overflowValue))
-            return false;
-
-        var value = overflowValue.Trim().ToLowerInvariant();
-        return value.Contains("hidden") || value.Contains("scroll") || value.Contains("auto") || value.Contains("clip");
-    }
-
-    private (double MinLeft, double MaxLeft, double MinTop, double MaxTop) GetScrollBounds(DomElement element)
-    {
-        var isRoot = IsViewportElementForMetrics(element);
-        var maxLeft = Math.Max(0, GetScrollWidthForDomElement(element, isRoot) - GetClientWidthForDomElement(element, isRoot));
-        var maxTop = Math.Max(0, GetScrollHeightForDomElement(element, isRoot) - GetClientHeightForDomElement(element, isRoot));
-
-        var props = GetComputedProps(element);
-        var writingMode = props.GetValueOrDefault("writing-mode")?.Trim().ToLowerInvariant();
-        var direction = props.GetValueOrDefault("direction");
-        var isRtl = string.Equals(direction, "rtl", StringComparison.OrdinalIgnoreCase);
-        var isVertical = IsVerticalWritingMode(writingMode);
-        var usesNegativeLeft = (isVertical && writingMode?.EndsWith("-rl", StringComparison.Ordinal) == true)
-            || (string.Equals(writingMode, "horizontal-tb", StringComparison.OrdinalIgnoreCase) && isRtl);
-        var usesNegativeTop = isVertical && isRtl;
-
-        var minLeft = usesNegativeLeft ? -maxLeft : 0;
-        var boundedMaxLeft = usesNegativeLeft ? 0 : maxLeft;
-        var minTop = usesNegativeTop ? -maxTop : 0;
-        var boundedMaxTop = usesNegativeTop ? 0 : maxTop;
-        return (minLeft, boundedMaxLeft, minTop, boundedMaxTop);
-    }
-
-    private bool CanProgrammaticallyScrollSelectListBox(DomElement element, bool vertical)
-    {
-        var props = GetComputedProps(element);
-        bool verticalWritingMode = IsVerticalWritingMode(props.GetValueOrDefault("writing-mode"));
-        bool blockAxisIsVertical = !verticalWritingMode;
-        if (vertical != blockAxisIsVertical)
-            return false;
-
-        double clientExtent = vertical ? GetClientHeightForDomElement(element, isRoot: false) : GetClientWidthForDomElement(element, isRoot: false);
-        double scrollExtent = vertical ? GetScrollHeightForDomElement(element, isRoot: false) : GetScrollWidthForDomElement(element, isRoot: false);
-        return scrollExtent > clientExtent + 0.5;
-    }
-
-    private bool TryGetSelectListBoxScrollExtent(DomElement element, bool verticalAxis, out double extent)
-    {
-        if (!IsSelectListBox(element))
-        {
-            extent = 0;
-            return false;
-        }
-
-        var props = GetComputedProps(element);
-        bool verticalWritingMode = IsVerticalWritingMode(props.GetValueOrDefault("writing-mode"));
-        int optionCount = Math.Max(1, CountSelectOptions(element));
-        double rowExtent = Math.Max(16, ResolveLineHeightForElement(element));
-        double clientInlineExtent = verticalWritingMode
-            ? GetClientHeightForDomElement(element, isRoot: false)
-            : GetClientWidthForDomElement(element, isRoot: false);
-        double clientBlockExtent = verticalWritingMode
-            ? GetClientWidthForDomElement(element, isRoot: false)
-            : GetClientHeightForDomElement(element, isRoot: false);
-        double totalBlockExtent = Math.Max(clientBlockExtent, optionCount * rowExtent);
-
-        extent = verticalAxis
-            ? (verticalWritingMode ? clientInlineExtent : totalBlockExtent)
-            : (verticalWritingMode ? totalBlockExtent : clientInlineExtent);
-        return true;
-    }
-
-    private static int CountSelectOptions(DomElement element)
-    {
-        int count = 0;
-        foreach (var child in ChildElements(element).Where(c => !IsText(c)))
-        {
-            if (string.Equals(child.TagName, "option", StringComparison.OrdinalIgnoreCase))
-            {
-                count++;
-                continue;
-            }
-
-            count += CountSelectOptions(child);
-        }
-
-        return count;
-    }
-
-    private static JSObject CreateSvgLengthValue(double numericValue)
-    {
-        var svgLength = new JSObject();
-        svgLength.FastAddValue((KeyString)"value", new JSNumber(numericValue), JSPropertyAttributes.EnumerableConfigurableValue);
-        svgLength.FastAddValue((KeyString)"valueInSpecifiedUnits", new JSNumber(numericValue), JSPropertyAttributes.EnumerableConfigurableValue);
-        svgLength.FastAddValue((KeyString)"unitType", new JSNumber(1), JSPropertyAttributes.EnumerableConfigurableValue);
-        svgLength.FastAddValue((KeyString)"SVG_LENGTHTYPE_UNKNOWN", new JSNumber(0), JSPropertyAttributes.EnumerableConfigurableValue);
-        svgLength.FastAddValue((KeyString)"SVG_LENGTHTYPE_NUMBER", new JSNumber(1), JSPropertyAttributes.EnumerableConfigurableValue);
-        svgLength.FastAddValue((KeyString)"SVG_LENGTHTYPE_PERCENTAGE", new JSNumber(2), JSPropertyAttributes.EnumerableConfigurableValue);
-        svgLength.FastAddValue((KeyString)"SVG_LENGTHTYPE_EMS", new JSNumber(3), JSPropertyAttributes.EnumerableConfigurableValue);
-        svgLength.FastAddValue((KeyString)"SVG_LENGTHTYPE_EXS", new JSNumber(4), JSPropertyAttributes.EnumerableConfigurableValue);
-        svgLength.FastAddValue((KeyString)"SVG_LENGTHTYPE_PX", new JSNumber(5), JSPropertyAttributes.EnumerableConfigurableValue);
-        svgLength.FastAddValue((KeyString)"SVG_LENGTHTYPE_CM", new JSNumber(6), JSPropertyAttributes.EnumerableConfigurableValue);
-        svgLength.FastAddValue((KeyString)"SVG_LENGTHTYPE_MM", new JSNumber(7), JSPropertyAttributes.EnumerableConfigurableValue);
-        svgLength.FastAddValue((KeyString)"SVG_LENGTHTYPE_IN", new JSNumber(8), JSPropertyAttributes.EnumerableConfigurableValue);
-        svgLength.FastAddValue((KeyString)"SVG_LENGTHTYPE_PT", new JSNumber(9), JSPropertyAttributes.EnumerableConfigurableValue);
-        svgLength.FastAddValue((KeyString)"SVG_LENGTHTYPE_PC", new JSNumber(10), JSPropertyAttributes.EnumerableConfigurableValue);
-        return svgLength;
-    }
-
-    private IEnumerable<DomElement> EnumerateRenderedDescendants(DomElement element)
-    {
-        foreach (var child in EnumerateRenderedChildren(element))
-        {
-            yield return child;
-
-            foreach (var descendant in EnumerateRenderedDescendants(child))
-                yield return descendant;
-        }
-    }
-
-    private IEnumerable<DomElement> EnumerateRenderedChildren(DomElement element)
-    {
-        if (string.Equals(element.TagName, "slot", StringComparison.OrdinalIgnoreCase))
-        {
-            var host = GetSlotHost(element);
-            if (host == null)
-                yield break;
-
-            foreach (var child in ChildElements(host))
-            {
-                if (!IsText(child) && SlotAcceptsNode(element, child))
-                    yield return child;
-            }
-
-            yield break;
-        }
-
-        foreach (var child in ChildElements(element))
-        {
-            if (!IsText(child))
-                yield return child;
-        }
-    }
-
-    private DomElement? FindScrollContainer(DomElement element)
-    {
-        var documentElement = GetOwningDocumentElement(element);
-        for (var current = GetScrollTraversalParent(element); current != null; current = GetScrollTraversalParent(current))
-        {
-            if (ReferenceEquals(current, documentElement))
-                return documentElement;
-
-            if (IsViewportBodyElement(current, documentElement))
-                continue;
-
-            var props = GetComputedProps(current);
-            if (HasOverflowClipping(props))
-                return current;
-        }
-
-        return documentElement;
-    }
-
-    private bool HasFixedPositionAncestorBefore(DomElement element, DomElement ancestor)
-    {
-        for (var current = GetScrollTraversalParent(element);
-             current != null && !ReferenceEquals(current, ancestor);
-             current = GetScrollTraversalParent(current))
-        {
-            var props = GetComputedProps(current);
-            if (string.Equals(props.GetValueOrDefault("position"), "fixed", StringComparison.OrdinalIgnoreCase))
-                return true;
-        }
-
-        return false;
-    }
-
-    private static bool IsDocumentElement(DomElement element) =>
-        string.Equals(element.TagName, "html", StringComparison.OrdinalIgnoreCase);
-
-    private bool IsViewportElementForMetrics(DomElement element)
-    {
-        var documentElement = GetOwningDocumentElement(element);
-        return IsDocumentElement(element) || IsViewportBodyElement(element, documentElement);
-    }
-
-    private static bool IsViewportBodyElement(DomElement element, DomElement documentElement) =>
-        string.Equals(element.TagName, "body", StringComparison.OrdinalIgnoreCase) &&
-        ReferenceEquals(ParentEl(element), documentElement);
-
-    private DomElement GetOwningDocumentElement(DomElement element)
-    {
-        for (var current = element; current != null; current = ParentEl(current)!)
-        {
-            if (string.Equals(current.TagName, "html", StringComparison.OrdinalIgnoreCase))
-                return current;
-        }
-
-        return DocumentElement;
-    }
-
-    private bool HasFixedPositionInDocument(DomElement element, DomElement documentElement)
-    {
-        if (IsFixedPositionElement(element))
-            return true;
-
-        return HasFixedPositionAncestorBefore(element, documentElement);
-    }
-
-    private bool IsFixedPositionElement(DomElement element)
-    {
-        var props = GetComputedProps(element);
-        return string.Equals(props.GetValueOrDefault("position"), "fixed", StringComparison.OrdinalIgnoreCase);
-    }
-
-    private DomElement? GetOuterFrameElement(DomElement documentElement)
-    {
-        // A sub-document's documentElement (<html>) now hangs off its canonical DomDocument
-        // (the severed content document, P4.4b); recover the owning frame via the reverse map
-        // (was ParentEl(ParentEl(<html>)) through the #subdoc-root element).
-        return GetFrameForContentDocument(documentElement?.ParentNode);
-    }
-
-    private DomElement? GetOuterScrollContinuationElement(DomElement scrollContainer)
-    {
-        if (IsDocumentElement(scrollContainer))
-            return GetOuterFrameElement(scrollContainer);
-
-        return scrollContainer;
-    }
-
-    /// <summary>
-    /// RF-BRIDGE-1b: the scroll-aware offset of <paramref name="element"/> within
-    /// <paramref name="ancestor"/> from the shared snapshot. The snapshot is a *natural*
-    /// (unscrolled) layout, so the element-border-to-ancestor-padding delta it yields is
-    /// the pre-scroll offset; the JS-set scroll offset of each intermediate scroll
-    /// container strictly between <paramref name="element"/> and <paramref name="ancestor"/>
-    /// is then subtracted (the ancestor's own scroll is left to the caller). Returns
-    /// <c>false</c> — so the caller reports zero — when either box is missing, or zoom is in
-    /// play (a zoomed cross-ancestor delta is not reconciled against the baked snapshot).
-    /// </summary>
-    private bool TrySharedOffsetWithinAncestor(DomElement element, DomElement ancestor, bool vertical, out double offset,
-        bool viewportAnchored = false)
-    {
-        offset = 0;
-        if (!UseSharedLayoutGeometry)
-            return false;
-        // (RF-BRIDGE-1b Track 3.2) The former cross-frame gate is gone: the shared
-        // snapshot now composes each subframe's geometry into the main coordinate frame
-        // (CssBox.LayoutNestedBrowsingContexts) *and* a position:fixed / root-anchored
-        // abspos element inside a subframe resolves against the frame's own sub-viewport
-        // (CssBox.IsNestedViewportRoot / FixedPositioningViewport), so a cross-frame
-        // element's composed box is correct. When the subframe box is absent from the
-        // snapshot (e.g. a cross-origin or non-materialised frame), the
-        // TryGetSharedLayoutGeometry lookups below return false and the caller still
-        // falls back to the estimator's frame-aware walk.
-        // (RF-BRIDGE-1b Track 3.1) The former abspos-in-inline-CB bypass is gone: the
-        // layout engine now places an absolutely/fixed-positioned element whose
-        // containing block is an inline box at its inset position, so its shared box is
-        // correct and the estimator fallback is no longer needed here.
-        if (!TryGetSharedLayoutGeometry(element, out var elementBox) ||
-            !TryGetSharedLayoutGeometry(ancestor, out var ancestorBox))
-            return false;
-
-        // Both edges are in the snapshot's zoom-baked document space; the offset is expressed
-        // in the ANCESTOR's coordinate frame, so divide the delta by the ancestor's cumulative
-        // used zoom to recover unzoomed CSS pixels. (Dividing by the *element's* own zoom would
-        // be wrong when only the element is zoomed: its own zoom scales its size/content, not
-        // its flow position within the ancestor.) A no-op when the ancestor is unzoomed. The
-        // former zoom gate that deferred zoomed elements to the estimator is gone, so the
-        // estimator can be deleted. Intermediate scroll offsets are already in unzoomed CSS
-        // pixels (JS scrollTo), so they subtract cleanly after this division.
-        var offsetZoom = GetUsedZoomForElement(ancestor);
-        var inverseOffsetZoom = offsetZoom > 0.0001 ? 1.0 / offsetZoom : 1.0;
-        double natural = (vertical
-            ? elementBox.BorderBox.Top - ancestorBox.PaddingBox.Top
-            : elementBox.BorderBox.Left - ancestorBox.PaddingBox.Left) * inverseOffsetZoom;
-
-        // (RF-BRIDGE-1b Track 3.3) For a viewport-anchored target — one being scrolled
-        // into the *visual* viewport whose containing chain includes a position:fixed
-        // ancestor F — only the scroll containers at or below F move the target: the
-        // target rides F's own overflow scroll (and any scroller nested inside F), but
-        // F itself is pinned to the viewport, so scroll containers *above* F do not
-        // shift it. Stop subtracting once the walk climbs past F. (A normal, non-anchored
-        // target subtracts every intermediate scroll container up to the ancestor.)
-        var fixedAnchor = viewportAnchored ? FindNearestFixedAncestorOrSelf(element) : null;
-
-        for (var current = element; ;)
-        {
-            var parent = GetScrollTraversalParent(current);
-            if (parent == null || ReferenceEquals(parent, ancestor))
-                break;
-            if (fixedAnchor != null && !IsDomDescendantOrSelf(parent, fixedAnchor))
-                break;
-            natural -= GetElementScrollOffset(parent, vertical);
-            current = parent;
-        }
-
-        offset = natural;
-        return true;
-    }
-
-    /// <summary>
-    /// Walks the DOM ancestry from <paramref name="element"/> (inclusive) and returns the
-    /// nearest <c>position:fixed</c> box, or null if none. Used to bound the scroll
-    /// subtraction for a viewport-anchored scrollIntoView target
-    /// (<see cref="TrySharedOffsetWithinAncestor"/>).
-    /// </summary>
-    private DomElement? FindNearestFixedAncestorOrSelf(DomElement element)
-    {
-        for (var current = element; current != null; current = ParentEl(current))
-        {
-            if (IsFixedPositionElement(current))
-                return current;
-        }
-        return null;
-    }
-
-    private static bool IsDomDescendantOrSelf(DomElement node, DomElement potentialAncestor)
-    {
-        for (var current = node; current != null; current = ParentEl(current))
-        {
-            if (ReferenceEquals(current, potentialAncestor))
-                return true;
-        }
-        return false;
-    }
-
-    /// <summary>
-    /// RF-BRIDGE-1b Track 3.3: offset of a viewport-anchored (fixed-subtree) target
-    /// within <paramref name="ancestor"/> from the shared snapshot, subtracting only the
-    /// scroll of containers at or below the target's nearest fixed ancestor
-    /// (<see cref="TrySharedOffsetWithinAncestor"/> with <c>viewportAnchored</c>), else
-    /// the estimator. Used by the visual-viewport fixed-element scrollIntoView sites so
-    /// they no longer call the estimator directly.
-    /// </summary>
-    private double OffsetWithinAncestorForFixedPreferShared(DomElement element, DomElement ancestor, bool vertical) =>
-        TrySharedOffsetWithinAncestor(element, ancestor, vertical, out var shared, viewportAnchored: true)
-            ? shared
-            : 0;
-
-    /// <summary>
-    /// RF-BRIDGE-1b: offset of <paramref name="element"/> within
-    /// <paramref name="ancestor"/> from the scroll-aware shared snapshot
-    /// (<see cref="TrySharedOffsetWithinAncestor"/>). With the coarse estimators deleted, a
-    /// shared-unavailable element (cross-origin / non-materialised frame) reports 0 — real
-    /// in-flow sticky/scrollIntoView targets are always present in the snapshot.
-    /// </summary>
-    private double OffsetWithinAncestorPreferShared(DomElement element, DomElement ancestor, bool vertical) =>
-        TrySharedOffsetWithinAncestor(element, ancestor, vertical, out var shared)
-            ? shared
-            : 0;
-
-    private double ResolveScrollIntoViewOffset(DomElement element, DomElement scrollContainer,
-        bool vertical, string? alignment,
-        double? viewportSizeOverride = null,
-        double? currentScrollOverride = null,
-        double? offsetOverride = null,
-        bool coordinateSpaceIsPhysical = false)
-    {
-        var normalizedAlignment = NormalizeScrollIntoViewAlignment(alignment, "start");
-        var offset = offsetOverride ?? OffsetWithinAncestorPreferShared(element, scrollContainer, vertical);
-        var targetSize = vertical ? GetBorderBoxHeight(GetComputedProps(element), element) : GetBorderBoxWidth(GetComputedProps(element), element);
-        var (marginStart, marginStartOwner) = ResolveScrollIntoViewInset(element, vertical ? "scroll-margin-top" : "scroll-margin-left");
-        var (marginEnd, marginEndOwner) = ResolveScrollIntoViewInset(element, vertical ? "scroll-margin-bottom" : "scroll-margin-right");
-        var (paddingStart, paddingStartOwner) = ResolveScrollIntoViewInset(scrollContainer, vertical ? "scroll-padding-top" : "scroll-padding-left");
-        var (paddingEnd, paddingEndOwner) = ResolveScrollIntoViewInset(scrollContainer, vertical ? "scroll-padding-bottom" : "scroll-padding-right");
-        marginStart = ConvertInsetToScrollContainerCoordinates(marginStart, marginStartOwner, scrollContainer);
-        marginEnd = ConvertInsetToScrollContainerCoordinates(marginEnd, marginEndOwner, scrollContainer);
-        paddingStart = ConvertInsetToScrollContainerCoordinates(paddingStart, paddingStartOwner, scrollContainer);
-        paddingEnd = ConvertInsetToScrollContainerCoordinates(paddingEnd, paddingEndOwner, scrollContainer);
-        var viewportSize = viewportSizeOverride ?? (vertical
-            ? GetClientHeightForDomElement(scrollContainer, IsDocumentElement(scrollContainer))
-            : GetClientWidthForDomElement(scrollContainer, IsDocumentElement(scrollContainer)));
-        var currentScroll = currentScrollOverride ?? GetElementScrollOffset(scrollContainer, vertical);
-        var physicalCurrentScroll = coordinateSpaceIsPhysical
-            ? currentScroll
-            : ConvertScrollCoordinateToPhysicalPosition(scrollContainer, vertical, currentScroll);
-
-        var startTarget = offset - marginStart - paddingStart;
-        var endTarget = offset + targetSize + marginEnd + paddingEnd - viewportSize;
-        if (normalizedAlignment == "start")
-            return ConvertPhysicalScrollPosition(scrollContainer, vertical, startTarget, coordinateSpaceIsPhysical);
-        if (normalizedAlignment == "end")
-            return ConvertPhysicalScrollPosition(scrollContainer, vertical, endTarget, coordinateSpaceIsPhysical);
-
-        var alignmentViewportSize = Math.Max(0, viewportSize - paddingStart - paddingEnd);
-        if (normalizedAlignment == "center")
-        {
-            var targetCenter = offset + ((targetSize + marginEnd - marginStart) / 2.0);
-            return ConvertPhysicalScrollPosition(
-                scrollContainer,
-                vertical,
-                targetCenter - paddingStart - (alignmentViewportSize / 2.0),
-                coordinateSpaceIsPhysical);
-        }
-
-        var visibleStart = physicalCurrentScroll + paddingStart;
-        var visibleEnd = physicalCurrentScroll + viewportSize - paddingEnd;
-        var targetStart = offset - marginStart;
-        var targetEnd = offset + targetSize + marginEnd;
-
-        if (targetStart >= visibleStart && targetEnd <= visibleEnd)
-            return currentScroll;
-
-        if (normalizedAlignment == "start-if-needed")
-            return ConvertPhysicalScrollPosition(scrollContainer, vertical, startTarget, coordinateSpaceIsPhysical);
-        if (normalizedAlignment == "end-if-needed")
-            return ConvertPhysicalScrollPosition(scrollContainer, vertical, endTarget, coordinateSpaceIsPhysical);
-
-        if (targetSize + marginStart + marginEnd > alignmentViewportSize)
-        {
-            var chosenTarget = Math.Abs(startTarget - physicalCurrentScroll) <= Math.Abs(endTarget - physicalCurrentScroll)
-                ? startTarget
-                : endTarget;
-            return ConvertPhysicalScrollPosition(scrollContainer, vertical, chosenTarget, coordinateSpaceIsPhysical);
-        }
-
-        if (targetStart < visibleStart)
-            return ConvertPhysicalScrollPosition(scrollContainer, vertical, startTarget, coordinateSpaceIsPhysical);
-        if (targetEnd > visibleEnd)
-            return ConvertPhysicalScrollPosition(scrollContainer, vertical, endTarget, coordinateSpaceIsPhysical);
-
-        return currentScroll;
-    }
-
-    private double ConvertPhysicalScrollPosition(DomElement scrollContainer,
-        bool vertical, double physicalPosition, bool coordinateSpaceIsPhysical)
-        => coordinateSpaceIsPhysical
-            ? physicalPosition
-            : ConvertPhysicalPositionToScrollCoordinate(scrollContainer, vertical, physicalPosition);
-
-    private double ConvertScrollCoordinateToPhysicalPosition(DomElement scrollContainer, bool vertical, double coordinate)
-    {
-        if (!UsesNegativeScrollCoordinates(scrollContainer, vertical))
-            return coordinate;
-
-        return coordinate + GetPositiveScrollExtent(scrollContainer, vertical);
-    }
-
-    private double ConvertPhysicalPositionToScrollCoordinate(DomElement scrollContainer, bool vertical, double physicalPosition)
-    {
-        if (!UsesNegativeScrollCoordinates(scrollContainer, vertical))
-            return physicalPosition;
-
-        return physicalPosition - GetPositiveScrollExtent(scrollContainer, vertical);
-    }
-
-    private bool UsesNegativeScrollCoordinates(DomElement scrollContainer, bool vertical)
-    {
-        var (minLeft, _, minTop, _) = GetScrollBounds(scrollContainer);
-        return vertical ? minTop < 0 : minLeft < 0;
-    }
-
-    private double GetPositiveScrollExtent(DomElement scrollContainer, bool vertical)
-    {
-        var (minLeft, maxLeft, minTop, maxTop) = GetScrollBounds(scrollContainer);
-        return vertical
-            ? (minTop < 0 ? maxTop - minTop : maxTop)
-            : (minLeft < 0 ? maxLeft - minLeft : maxLeft);
-    }
-
-    private (double Value, DomElement Owner) ResolveScrollIntoViewInset(DomElement element, string propertyName)
-    {
-        var props = GetComputedProps(element);
-        var value = props.GetValueOrDefault(propertyName);
-        if (string.Equals(value, "inherit", StringComparison.OrdinalIgnoreCase) && ParentEl(element) != null)
-            return ResolveScrollIntoViewInset(ParentEl(element), propertyName);
-
-        return (ParseCssLengthToPixelsWithViewport(value, element), element);
-    }
-
-    private double ConvertInsetToScrollContainerCoordinates(double inset, DomElement insetOwner, DomElement scrollContainer)
-    {
-        if (!double.IsFinite(inset) || AreClose(inset, 0))
-            return 0;
-
-        var ownerZoom = GetUsedZoomForElement(insetOwner);
-        var containerZoom = GetUsedZoomForElement(scrollContainer);
-        if (!double.IsFinite(ownerZoom) || ownerZoom <= 0 ||
-            !double.IsFinite(containerZoom) || containerZoom <= 0)
-        {
-            return inset;
-        }
-
-        return inset * (ownerZoom / containerZoom);
-    }
-
-    private double ParseCssLengthToPixelsWithViewport(string? value, DomElement? referenceElement = null,
-        bool forLineHeight = false, double? percentageBasis = null, bool forFontSize = false)
-    {
-        if (string.IsNullOrWhiteSpace(value))
-            return 0;
-
-        return TryEvaluateCssLengthWithViewport(value, referenceElement, forLineHeight, percentageBasis, out var px, forFontSize)
-            ? px
-            : 0;
-    }
-
-    private bool TryEvaluateCssLengthWithViewport(
-        string value,
-        DomElement? referenceElement,
-        bool forLineHeight,
-        double? percentageBasis,
-        out double result,
-        bool forFontSize = false)
-    {
-        result = 0;
-        if (string.IsNullOrWhiteSpace(value))
-            return false;
-
-        var normalized = value.Trim();
-        while (normalized.Length >= 2 &&
-               normalized[0] == '(' &&
-               normalized[^1] == ')' &&
-               HasBalancedParens(normalized[1..^1]))
-        {
-            normalized = normalized[1..^1].Trim();
-        }
-
-        if (TryEvaluateMathLengthFunction(normalized, referenceElement, forLineHeight, percentageBasis, out result, forFontSize))
-            return true;
-
-        var additiveOperatorIndex = FindTopLevelAdditiveOperator(normalized);
-        if (additiveOperatorIndex > 0)
-        {
-            if (!TryEvaluateCssLengthWithViewport(
-                    normalized[..additiveOperatorIndex],
-                    referenceElement,
-                    forLineHeight,
-                    percentageBasis,
-                    out var left,
-                    forFontSize) ||
-                !TryEvaluateCssLengthWithViewport(
-                    normalized[(additiveOperatorIndex + 1)..],
-                    referenceElement,
-                    forLineHeight,
-                    percentageBasis,
-                    out var right,
-                    forFontSize))
-            {
-                return false;
-            }
-
-            result = normalized[additiveOperatorIndex] == '+'
-                ? left + right
-                : left - right;
-            return true;
-        }
-
-        var lower = normalized.ToLowerInvariant();
-        if (percentageBasis.HasValue && lower.EndsWith('%') &&
-            double.TryParse(lower[..^1], NumberStyles.Float,
-                CultureInfo.InvariantCulture, out var percent))
-        {
-            result = percentageBasis.Value * (percent / 100.0);
-            return true;
-        }
-
-        if (referenceElement != null &&
-            lower.EndsWith("rem") &&
-            double.TryParse(lower[..^3], NumberStyles.Float,
-                CultureInfo.InvariantCulture, out var rem))
-        {
-            result = rem * ResolveFontSizeForLength(referenceElement, rootRelative: true);
-            return true;
-        }
-
-        if (referenceElement != null &&
-            lower.EndsWith("em") &&
-            double.TryParse(lower[..^2], NumberStyles.Float,
-                CultureInfo.InvariantCulture, out var em))
-        {
-            // For the font-size property itself, em resolves against the parent's
-            // font-size (not the element's own), otherwise resolving the element's
-            // font-size would recurse into itself.
-            double emBasis;
-            if (forFontSize)
-            {
-                var parent = ParentEl(referenceElement);
-                emBasis = parent != null ? ResolveFontSizeForElement(parent) : 16;
-            }
-            else
-            {
-                emBasis = ResolveFontSizeForLength(referenceElement, rootRelative: false);
-            }
-
-            result = em * emBasis;
-            return true;
-        }
-
-        if (referenceElement != null &&
-            lower.EndsWith("rlh") &&
-            double.TryParse(lower[..^3], NumberStyles.Float,
-                CultureInfo.InvariantCulture, out var rlh))
-        {
-            result = rlh * ResolveLineHeightForLength(referenceElement, rootRelative: true);
-            return true;
-        }
-
-        if (referenceElement != null &&
-            lower.EndsWith("lh") &&
-            double.TryParse(lower[..^2], NumberStyles.Float,
-                CultureInfo.InvariantCulture, out var lh))
-        {
-            result = lh * ResolveLineHeightForLength(referenceElement, rootRelative: false, forLineHeight);
-            return true;
-        }
-
-        var px = ParseCssLengthToPixels(normalized, _viewportWidth, _viewportHeight);
-        if (double.IsNaN(px))
-            return false;
-
-        result = px;
-        return true;
-    }
-
-    private bool TryEvaluateMathLengthFunction(string value, DomElement? referenceElement,
-        bool forLineHeight, double? percentageBasis, out double result,
-        bool forFontSize = false)
-    {
-        result = 0;
-        if (string.IsNullOrWhiteSpace(value) || value[^1] != ')')
-            return false;
-
-        static bool StartsWithFunction(string candidate, string functionName)
-            => candidate.StartsWith(functionName + "(", StringComparison.OrdinalIgnoreCase);
-
-        if (StartsWithFunction(value, "calc"))
-        {
-            var content = value[5..^1];
-            return HasBalancedParens(content) &&
-                   TryEvaluateCssLengthWithViewport(content, referenceElement, forLineHeight, percentageBasis, out result, forFontSize);
-        }
-
-        if (!StartsWithFunction(value, "min") && !StartsWithFunction(value, "max"))
-            return false;
-
-        var isMax = StartsWithFunction(value, "max");
-        var contentValue = value[4..^1];
-        if (!HasBalancedParens(contentValue))
-            return false;
-
-        var parts = SplitTopLevelArguments(contentValue);
-        if (parts.Count == 0)
-            return false;
-
-        double? candidate = null;
-        foreach (var part in parts)
-        {
-            if (!TryEvaluateCssLengthWithViewport(part, referenceElement, forLineHeight, percentageBasis, out var parsed, forFontSize))
-                return false;
-
-            candidate = candidate.HasValue
-                ? (isMax ? Math.Max(candidate.Value, parsed) : Math.Min(candidate.Value, parsed))
-                : parsed;
-        }
-
-        if (!candidate.HasValue)
-            return false;
-
-        result = candidate.Value;
-        return true;
-    }
-
-    private static int FindTopLevelAdditiveOperator(string expression)
-    {
-        var depth = 0;
-        for (int i = expression.Length - 1; i >= 1; i--)
-        {
-            switch (expression[i])
-            {
-                case ')':
-                    depth++;
-                    break;
-                case '(':
-                    depth--;
-                    break;
-                case '+':
-                case '-':
-                    if (depth != 0)
-                        break;
-
-                    var leftIndex = i - 1;
-                    while (leftIndex >= 0 && char.IsWhiteSpace(expression[leftIndex]))
-                        leftIndex--;
-
-                    var rightIndex = i + 1;
-                    while (rightIndex < expression.Length && char.IsWhiteSpace(expression[rightIndex]))
-                        rightIndex++;
-
-                    if (leftIndex >= 0 &&
-                        rightIndex < expression.Length &&
-                        expression[leftIndex] != '(' &&
-                        expression[leftIndex] != ',' &&
-                        expression[leftIndex] != '+' &&
-                        expression[leftIndex] != '-')
-                    {
-                        return i;
-                    }
-                    break;
-            }
-        }
-
-        return -1;
-    }
-
-    private static List<string> SplitTopLevelArguments(string value)
-    {
-        var parts = new List<string>();
-        var depth = 0;
-        var start = 0;
-
-        for (int i = 0; i < value.Length; i++)
-        {
-            switch (value[i])
-            {
-                case '(':
-                    depth++;
-                    break;
-                case ')':
-                    depth--;
-                    if (depth < 0)
-                        return [];
-                    break;
-                case ',' when depth == 0:
-                    parts.Add(value[start..i].Trim());
-                    start = i + 1;
-                    break;
-            }
-        }
-
-        if (depth != 0)
-            return [];
-
-        parts.Add(value[start..].Trim());
-        return parts;
-    }
-
-    private double ResolveContainingBlockReferenceLength(DomElement element, bool vertical)
-    {
-        if (ParentEl(element) == null ||
-            string.Equals(element.TagName, "html", StringComparison.OrdinalIgnoreCase) ||
-            string.Equals(element.TagName, "body", StringComparison.OrdinalIgnoreCase) ||
-            string.Equals(ParentEl(element).TagName, "html", StringComparison.OrdinalIgnoreCase) ||
-            string.Equals(ParentEl(element).TagName, "body", StringComparison.OrdinalIgnoreCase))
-        {
-            return GetViewportReferenceLength(element, vertical);
-        }
-
-        var (Left, Top, Width, Height) = ComputeUnzoomedLayoutRect(ParentEl(element));
-        var reference = vertical ? Height : Width;
-        return reference > 0 ? reference : GetViewportReferenceLength(element, vertical);
-    }
-
-    private double GetViewportReferenceLength(DomElement? element, bool vertical)
-    {
-        if (element != null)
-        {
-            var documentElement = GetOwningDocumentElement(element);
-            var frameElement = GetOuterFrameElement(documentElement);
-            if (frameElement != null)
-            {
-                var frameProps = GetComputedProps(frameElement);
-                var frameLength = ParseCssLengthToPixelsWithViewport(
-                    frameProps.GetValueOrDefault(vertical ? "height" : "width"),
-                    frameElement);
-                if (frameLength > 0)
-                    return frameLength;
-
-                if (TryGetAttribute(frameElement, vertical ? "height" : "width", out var frameAttribute) &&
-                    double.TryParse(frameAttribute, NumberStyles.Float,
-                        CultureInfo.InvariantCulture, out frameLength) &&
-                    frameLength > 0)
-                {
-                    return frameLength;
-                }
-            }
-        }
-
-        return vertical ? _viewportHeight : _viewportWidth;
-    }
-
-    private static bool HasExplicitBodyMargin(string? value)
-    {
-        return !string.IsNullOrWhiteSpace(value) &&
-               !string.Equals(value, "auto", StringComparison.OrdinalIgnoreCase);
-    }
-
-    private double ResolveLineHeightForLength(DomElement element, bool rootRelative, bool forLineHeight = false)
-    {
-        var target = rootRelative ? GetRootElement(element) : (forLineHeight ? ParentEl(element) ?? element : element);
-        return ResolveLineHeightForElement(target);
-    }
-
-    private double ResolveFontSizeForLength(DomElement element, bool rootRelative)
-    {
-        var target = rootRelative ? GetRootElement(element) : element;
-        return ResolveFontSizeForElement(target);
-    }
-
-    private DomElement GetRootElement(DomElement element)
-    {
-        DomElement? htmlElement = null;
-        var current = element;
-        while (ParentEl(current) != null)
-        {
-            current = ParentEl(current);
-            if (string.Equals(current.TagName, "html", StringComparison.OrdinalIgnoreCase))
-                htmlElement = current;
-        }
-
-        return htmlElement ?? current;
-    }
-
-    private double ResolveLineHeightForElement(DomElement element)
-    {
-        var props = GetComputedProps(element);
-        var fontSize = ResolveFontSizeForElement(element);
-        var lineHeight = props.GetValueOrDefault("line-height");
-        if (string.IsNullOrWhiteSpace(lineHeight) ||
-            string.Equals(lineHeight, "normal", StringComparison.OrdinalIgnoreCase))
-        {
-            return fontSize * 1.2;
-        }
-
-        var normalized = lineHeight.Trim().ToLowerInvariant();
-        if (double.TryParse(normalized, NumberStyles.Float,
-                CultureInfo.InvariantCulture, out var multiplier))
-        {
-            return fontSize * multiplier;
-        }
-
-        return ParseCssLengthToPixelsWithViewport(lineHeight, element, forLineHeight: true);
-    }
-
-    private double ResolveFontSizeForElement(DomElement element)
-    {
-        var props = GetComputedProps(element);
-        var fontSize = ParseCssLengthToPixelsWithViewport(props.GetValueOrDefault("font-size"), element, forFontSize: true);
-        if (fontSize > 0)
-            return fontSize;
-
-        for (var current = element; current != null; current = ParentEl(current))
-        {
-            if (!TryGetAttribute(current, "font-size", out var attributeValue) ||
-                string.IsNullOrWhiteSpace(attributeValue))
-            {
-                continue;
-            }
-
-            var attributeFontSize = ParseCssLengthToPixelsWithViewport(attributeValue, current, forFontSize: true);
-            if (attributeFontSize > 0)
-                return attributeFontSize;
-        }
-
-        return 16;
-    }
-
-    [GeneratedRegex(@"scale\(\s*(?<value>[-+]?[0-9]*\.?[0-9]+)\s*\)", RegexOptions.IgnoreCase, "de-DE")]
-    private static partial System.Text.RegularExpressions.Regex GetTransformScaleRegex();
-    [GeneratedRegex(@"[Mm]\s*(?<x>[-+]?[0-9]*\.?[0-9]+)(?:[\s,]+(?<y>[-+]?[0-9]*\.?[0-9]+))", RegexOptions.CultureInvariant)]
-    private static partial System.Text.RegularExpressions.Regex TryResolveSvgTextPathStartRegex();
 }
